+++ actors/src/test/kotlin/kts/actors/actionsTest.kt
@@ -2,9 +2,7 @@ package kts.actors
 
 import com.badlogic.gdx.scenes.scene2d.Action
 import com.badlogic.gdx.scenes.scene2d.Actor
-import com.badlogic.gdx.scenes.scene2d.actions.ParallelAction
 import com.badlogic.gdx.scenes.scene2d.actions.RepeatAction
-import com.badlogic.gdx.scenes.scene2d.actions.SequenceAction
 import ktx.actors.*
 import org.junit.Assert.*
 import org.junit.Test
@@ -66,7 +64,6 @@ class ActionsTest {
 
     val sequence = firstAction.then(secondAction) // === firstAction then secondAction
 
-    assertTrue(sequence is SequenceAction)
     assertEquals(firstAction, sequence.actions[0])
     assertEquals(secondAction, sequence.actions[1])
     assertEquals(2, sequence.actions.size)
@@ -81,7 +78,6 @@ class ActionsTest {
     // / Note that the second "then" is a different extension function - it prevents from creating multiple sequences.
     val sequence = firstAction then secondAction then thirdAction
 
-    assertTrue(sequence is SequenceAction)
     assertEquals(firstAction, sequence.actions[0])
     assertEquals(secondAction, sequence.actions[1])
     assertEquals(thirdAction, sequence.actions[2])
@@ -95,7 +91,6 @@ class ActionsTest {
 
     val parallel = firstAction.parallelTo(secondAction) // === firstAction parallelTo secondAction
 
-    assertTrue(parallel is ParallelAction)
     assertTrue(firstAction in parallel.actions)
     assertTrue(secondAction in parallel.actions)
     assertEquals(2, parallel.actions.size)
@@ -110,7 +105,6 @@ class ActionsTest {
     // Note that the second "parallelTo" is a different extension function - it prevents from creating multiple parallels.
     val parallel = firstAction parallelTo secondAction parallelTo thirdAction
 
-    assertTrue(parallel is ParallelAction)
     assertTrue(firstAction in parallel.actions)
     assertTrue(secondAction in parallel.actions)
     assertTrue(thirdAction in parallel.actions)
@@ -121,13 +115,12 @@ class ActionsTest {
   fun `should create actions repeating forever`() {
     val action = MockAction().repeatForever()
 
-    assertTrue(action is RepeatAction)
     assertEquals(RepeatAction.FOREVER, action.count)
   }
 
   /** Action testing utility. */
   class MockAction : Action() {
-    var acted = false
+    private var acted = false
     override fun act(delta: Float): Boolean {
       acted = true
       return true